+++ modules/apps/rate-limiting/types/errors.go
@@ -16,4 +16,8 @@ var (
 	ErrEpochNotFound          = errorsmod.Register(ModuleName, 9, "hour epoch not found in store")
 	ErrUnmarshalEpoch         = errorsmod.Register(ModuleName, 10, "could not unmarshal epochBz")
 	ErrInvalidEpoce           = errorsmod.Register(ModuleName, 11, "invalid hour epoch")
+	ErrInvalidChannelID       = errorsmod.Register(ModuleName, 12, "invalid channel")
+	ErrAsyncPacketNotFound    = errorsmod.Register(ModuleName, 13, "async packet not found")
+	ErrAsyncAckNil            = errorsmod.Register(ModuleName, 14, "async ack is nil")
+	ErrInvalidPacketData      = errorsmod.Register(ModuleName, 15, "invalid packet data")
 )